norme et encore de la norme
