+++ builtin/echo.c
@@ -6,7 +6,7 @@
 /*   By: cmontaig <cmontaig@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/05 16:43:03 by cmontaig          #+#    #+#             */
-/*   Updated: 2025/04/29 14:00:31 by cmontaig         ###   ########.fr       */
+/*   Updated: 2025/05/10 15:35:16 by cmontaig         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	print_echo(t_cmd *cmd)
 	t_token	*token;
 	int		newline;
 	bool	first;
-	
+
 	newline = 1;
 	first = true;
 	token = cmd->token->next;
@@ -30,7 +30,6 @@ void	print_echo(t_cmd *cmd)
 	{
 		if (!first)
 			write(STDOUT_FILENO, " ", 1);
-		
 		write(STDOUT_FILENO, token->value, ft_strlen(token->value));
 		first = false;
 		token = token->next;
